Activate example CI on Python3.9
This pull request only activates PyTorch example, i.e.
`pytorch_simple.py`.
The change to `setup.py` is quite tentative and far from what this
should look like, however, as of now this change is enough given the
number of requirements: it's cumbersome to add the constraints all the
other packages in 'example'.