Make default value of verbose_eval None to avoid warning
Since LightGBM 3.3.0, the following arguments are deprecated to use
callbacks instead:
- `verbose_eval`
- `early_stopping_rounds`
- `learning_rates`
- `eval_result`

However, the default value of `verbose_eval` for `LightGBMTuner` and
`LightGBMTunerCV` is `True` so that we need to write extra
`verbose_eval=None` to avoid blocking to use `log_evaluation` callback.

This patch avoids to write redundant `verbose_eval` argument for
`log_evaluation` callback.