Add `mock_before_trial` to check if cache clear happens at each trial
The original version test checks the cache clear after before_trial,
so if we call `_get_all_trials` with `use_cache=True` at `before_trial`,
this unit test fails.
To avoid this problem, we actually need to check the cache clear before
`before_trial`.
To this end, I introduced a mock function to do so.